+++ src/domain/communication/discussion/views/DiscussionsListView.tsx
@@ -4,7 +4,7 @@ import React, { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Filter } from '../../../platform/admin/components/Common/Filter';
 import ProfileCard from '../../../../common/components/composite/common/cards/ProfileCard/ProfileCard';
-import DiscussionOverview from '../../../../common/components/composite/entities/Communication/DiscussionOverview';
+import DiscussionOverview from './DiscussionOverview';
 import { Discussion } from '../models/Discussion';
 import { ViewProps } from '../../../../core/container/view';
