+++ src/domain/community/community/views/CommunitySection/DiscussionsView.tsx
@@ -4,7 +4,7 @@ import React, { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useResolvedPath } from 'react-router-dom';
 
-import DiscussionOverview from '../../../../../common/components/composite/entities/Communication/DiscussionOverview';
+import DiscussionOverview from '../../../../communication/discussion/views/DiscussionOverview';
 import ConditionalLink from '../../../../../common/components/core/ConditionalLink';
 import { RouterLink } from '../../../../../common/components/core/RouterLink';
 import { Discussion } from '../../../../communication/discussion/models/Discussion';
